Revised the implementation of methods in the ExportService classes to conform with SOLID principles and design patterns. This includes modifying the return types of certain functions (e.g., Context.DataProcess, PdfConverter.convertToPdf) and making changes to the PdfConverter and ZipFile classes. Additionally, logging functionality has been integrated.

+++ src/edu/najah/cap/data/exportservice/todownload/LocalDownload.java
@@ -1,7 +1,6 @@
 package edu.najah.cap.data.exportservice.todownload;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -9,9 +8,7 @@ import java.nio.file.StandardCopyOption;
 
 public class LocalDownload implements ILocalStorage {
     private static final Logger logger = LoggerFactory.getLogger(LocalDownload.class);
-
     private final String downloadDirectory;
-
     public LocalDownload(String downloadDirectory) {
         this.downloadDirectory = downloadDirectory;
     }
@@ -19,11 +16,10 @@ public class LocalDownload implements ILocalStorage {
     @Override
     public void downloadFile(String filePath) {
         File sourceFile = new File(filePath);
-//        if (!sourceFile.exists()) {
-//            logger.error("File does not exist: {}", filePath);
-//            return;
-//        }
-
+        if (!sourceFile.exists()) {
+            logger.error("File does not exist: {}", filePath);
+            return;
+        }
         File destinationFile = new File(downloadDirectory, sourceFile.getName());
         try {
             Files.copy(sourceFile.toPath(), destinationFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
